+++ src/components/ModalDialog.vue
@@ -67,7 +67,8 @@ watch(
         }
       })
     }
-  }
+  },
+  { immediate: true } // Needed to make async components work.
 )
 </script>