+++ src/components/modals/generation/EnumerateChord.vue
@@ -4,6 +4,10 @@ import { parseChordInput } from '@/utils'
 import { Scale } from 'scale-workshop-core'
 import { useModalStore } from '@/stores/modal'
 
+defineProps<{
+  show: boolean
+}>()
+
 const emit = defineEmits(['update:scale', 'update:scaleName', 'cancel'])
 
 const modal = useModalStore()
@@ -29,7 +33,7 @@ function generate() {
 </script>
 
 <template>
-  <Modal @confirm="generate" @cancel="$emit('cancel')">
+  <Modal :show="show" @confirm="generate" @cancel="$emit('cancel')">
     <template #header>
       <h2>Enumerate chord</h2>
     </template>