+++ src/components/modals/generation/EulerGenus.vue
@@ -5,6 +5,10 @@ import { clamp } from 'xen-dev-utils'
 import { Scale } from 'scale-workshop-core'
 import { useModalStore } from '@/stores/modal'
 
+defineProps<{
+  show: boolean
+}>()
+
 const emit = defineEmits(['update:scale', 'update:scaleName', 'cancel'])
 
 const modal = useModalStore()
@@ -22,7 +26,7 @@ function generate() {
 </script>
 
 <template>
-  <Modal @confirm="generate" @cancel="$emit('cancel')">
+  <Modal :show="show" @confirm="generate" @cancel="$emit('cancel')">
     <template #header>
       <h2>Generate Euler-Fokker genus</h2>
     </template>