+++ src/components/modals/generation/SpanLattice.vue
@@ -9,6 +9,7 @@ import { useLatticeStore } from '@/stores/tempering'
 import { setAndReportValidity } from '@/utils'
 
 const props = defineProps<{
+  show: boolean
   centsFractionDigits: number
 }>()
 
@@ -100,7 +101,7 @@ function generate() {
 </script>
 
 <template>
-  <Modal @confirm="generate" @cancel="$emit('cancel')">
+  <Modal :show="show" @confirm="generate" @cancel="$emit('cancel')">
     <template #header>
       <h2>Generate lattice</h2>
     </template>