+++ src/components/modals/modification/ConvertType.vue
@@ -6,6 +6,7 @@ import type { IntervalOptions, Scale } from 'scale-workshop-core'
 import { useModalStore } from '@/stores/modal'
 
 const props = defineProps<{
+  show: boolean
   scale: Scale
   centsFractionDigits: number
   decimalFractionDigits: number
@@ -49,7 +50,7 @@ function modify() {
 </script>
 
 <template>
-  <Modal @confirm="modify" @cancel="$emit('cancel')">
+  <Modal :show="show" @confirm="modify" @cancel="$emit('cancel')">
     <template #header>
       <h2>Convert interval values</h2>
     </template>