+++ src/components/modals/modification/TakeSubset.vue
@@ -5,6 +5,7 @@ import type { Scale } from 'scale-workshop-core'
 import { useModalStore } from '@/stores/modal'
 
 const props = defineProps<{
+  show: boolean
   scale: Scale
 }>()
 
@@ -38,7 +39,7 @@ function modify() {
 </script>
 
 <template>
-  <Modal @confirm="modify" @cancel="$emit('cancel')">
+  <Modal :show="show" @confirm="modify" @cancel="$emit('cancel')">
     <template #header>
       <h2>Take subset</h2>
     </template>